+++ src/extension/models/constants.ts
@@ -75,7 +75,7 @@ export enum Constants {
     NOTEBOOK_ID = 'dothttp-book',
 
     // download stuff
-    EXTENSION_VERSION = "0.0.32",
+    EXTENSION_VERSION = "0.0.33",
 
     dothttpNotebook = "dothttp-book",