+++ src/renderer/svg.d.ts
@@ -1,0 +1,4 @@
+declare module '*.svg' {
+    const OpenInEditor: string;
+    export = OpenInEditor;
+}